Move isaIgnoredImages
