+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/store/isa-groups.ts
@@ -1,7 +1,8 @@
 import { derived } from 'svelte/store';
 import { z } from 'zod';
 import { jetpack_boost_ds } from '../../../stores/data-sync-client';
-import { isaData, isaIgnoredImages } from './isa-data';
+import { isaData } from './isa-data';
+import { isaIgnoredImages } from './isa-ignored-images';
 
 const Group = z.object( {
 	name: z.string(),
