+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/store/isa-ignored-images.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod';
+import { jetpack_boost_ds } from '../../../stores/data-sync-client';
+import { ImageData } from './zod-types';
+
+const image_size_analysis_ignored_images = jetpack_boost_ds.createAsyncStore(
+	'image_size_analysis_ignored_images',
+	z.array( ImageData ).catch( [] )
+);
+
+export const isaIgnoredImages = image_size_analysis_ignored_images.store;
